resolve compiler warnings

+++ src/tests/test.c
@@ -16,15 +16,15 @@ void test_vector(void) {
     vector_t vec_int = vector_create_and_init(sizeof(int), 5);
     int* ints = vector_get_data(&vec_int);
 
-    for (int i = 0; i < vec_int.size; i++) {
-        ints[i] = i + 1;
+    for (size_t i = 0; i < vec_int.size; i++) {
+        ints[i] = (int)i + 1;
     }
     int twelve_k = 12000;
     vector_push_back(&vec_int, &twelve_k);
     vector_push_back(&vec_int, &twelve_k);
 
-    for (int i = 0; i < vec_int.size; i++) {
-        printf("idx: %d, val: %d\n", i, *(int*)vector_at(&vec_int, i));
+    for (size_t i = 0; i < vec_int.size; i++) {
+        printf("idx: %d, val: %d\n", (int)i, *(int*)vector_at(&vec_int, i));
     }
 
     vector_destroy(&vec_int);
